VIDEO: check for double fbo bind call

+++ src/modules/video/DepthBuffer.h
@@ -18,7 +18,7 @@ public:
 
 	void bind();
 	void unbind();
-	float *read();
+	uint8_t *read();
 
 	inline GLuint getTexture() const {
 		return _depthTexture;
